# Improve memory usage and perf of CodePointCharStream: Use 8-bit, 16-bit, or 32-bit buffer