… runtime-testsuite/test/org/antlr/v4/test/runtime/java/TestUTF8CodePointDecoder.java
… runtime/Java/src/org/antlr/v4/runtime/CodePointCharStream.java
… runtime/Java/src/org/antlr/v4/runtime/UTF8CodePointDecoder.java