This greatly improves the memory usage and performance of `CodePointCharStream` by ensuring the internal storage uses either an 8-bit buffer (for Unicode code points <= `U+00FF`), 16-bit buffer (for Unicode code points <= `U+FFFF`), or a 32-bit buffer (Unicode code points > `U+FFFF`).

I split out the internal storage into a class `CodePointBuffer` which has a `CodePointBuffer.Builder` class which has the logic to upgrade from 8-bit to 16-bit to 32-bit storage.

I found the perf hotspot in `CodePointCharStream` on master was the virtual method calls from `CharStream.LA(offset)` into `IntBuffer`.

Refactoring it into `CodePointBuffer` didn't help (in fact, it added another virtual method call).

To fix the perf, I made `CodePointCharStream` an `abstract` class and made three concrete subclasses: `CodePoint8BitCharStream`, `CodePoint16BitCharStream`, and `CodePoint32BitCharStream` which directly access the array of underlying code points in the `CodePointBuffer` without virtual method calls.