@parrt 2017-03-23
Cool.  Does the new lexer timing stuff say anything interesting about the cost of the new streams? I should probably add a memory size check to those tests... Java makes it hard but maybe there's a new API I don't know about.