@bhamiltoncx 2017-03-23
> Does the new lexer timing stuff say anything interesting about the cost of the new streams?

I'm trying to run the new test, but it looks like it passes a URL where we expect a Path:

```
% java -cp /data/users/beng/src/antlr4/runtime/Java/target/antlr4-runtime-4.7-SNAPSHOT.jar:/data/users/beng/src/antlr4/runtime-testsuite/target/antlr4-runtime-testsuite-4.7-SNAPSHOT.jar:/data/users/beng/src/antlr4/runtime-testsuite/target/antlr4-runtime-testsuite-4.7-SNAPSHOT-tests.jar org.antlr.v4.test.runtime.java.api.perf.TimeLexerSpeed      
Warming up Java compilerException in thread "main" java.nio.file.NoSuchFileException: file:/data/users/beng/src/antlr4/runtime-testsuite/target/antlr4-runtime-testsuite-4.7-SNAPSHOT.jar!/Java/src/org/antlr/v4/runtime/Parser.java
        at sun.nio.fs.UnixException.translateToIOException(UnixException.java:86)
        at sun.nio.fs.UnixException.rethrowAsIOException(UnixException.java:102)
        at sun.nio.fs.UnixException.rethrowAsIOException(UnixException.java:107)
        at sun.nio.fs.UnixFileSystemProvider.newByteChannel(UnixFileSystemProvider.java:214)
        at java.nio.file.Files.newByteChannel(Files.java:361)
        at java.nio.file.Files.newByteChannel(Files.java:407)
        at org.antlr.v4.runtime.CharStreams.fromPath(CharStreams.java:59)
        at org.antlr.v4.runtime.CharStreams.fromPath(CharStreams.java:43)
        at org.antlr.v4.test.runtime.java.api.perf.TimeLexerSpeed.lex_new_java_utf8(TimeLexerSpeed.java:208)
        at org.antlr.v4.test.runtime.java.api.perf.TimeLexerSpeed.compilerWarmUp(TimeLexerSpeed.java:110)
        at org.antlr.v4.test.runtime.java.api.perf.TimeLexerSpeed.main(TimeLexerSpeed.java:71)
```

I'll try to fix it.