@parrt 2017-03-23
ok, thanks. yeah, I've been running from the development environment not from the jar.  I should've known there was going to be an issue finding the files to test...