@bhamiltoncx 2017-03-23
Fixed the test: https://github.com/antlr/antlr4/pull/1787

Ran test on Linux with Intel Xeon E5-2600 @ 2.20 GHz.
                                    
Before:

```
Warming up Java compiler....                                                                                                    
    load_legacy_java_utf8 average time   248us over 3500 loads of 29038 symbols from Parser.java                                
    load_legacy_java_utf8 average time   301us over 3500 loads of 13379 symbols from udhr_hin.txt                               
            load_new_utf8 average time   535us over 3500 loads of 29038 symbols from Parser.java                                
            load_new_utf8 average time   420us over 3500 loads of 13379 symbols from udhr_hin.txt                               
                                                                                                                                
     lex_legacy_java_utf8 average time   625us over 2000 runs of 29038 symbols                                                  
     lex_legacy_java_utf8 average time  1678us over 2000 runs of 29038 symbols DFA cleared                                      
        lex_new_java_utf8 average time   644us over 2000 runs of 29038 symbols                                                  
        lex_new_java_utf8 average time  1534us over 2000 runs of 29038 symbols DFA cleared                                      
                                                                                                                                
 lex_legacy_grapheme_utf8 average time 12950us over  400 runs of  6614 symbols from udhr_kor.txt                                
 lex_legacy_grapheme_utf8 average time 11953us over  400 runs of  6614 symbols from udhr_kor.txt DFA cleared                    
 lex_legacy_grapheme_utf8 average time 10144us over  400 runs of 13379 symbols from udhr_hin.txt                                
 lex_legacy_grapheme_utf8 average time 11146us over  400 runs of 13379 symbols from udhr_hin.txt DFA cleared                    
    lex_new_grapheme_utf8 average time 11914us over  400 runs of  6614 symbols from udhr_kor.txt                                
    lex_new_grapheme_utf8 average time 12030us over  400 runs of  6614 symbols from udhr_kor.txt DFA cleared                    
    lex_new_grapheme_utf8 average time 10400us over  400 runs of 13379 symbols from udhr_hin.txt                                
    lex_new_grapheme_utf8 average time 10365us over  400 runs of 13379 symbols from udhr_hin.txt DFA cleared                    
    lex_new_grapheme_utf8 average time   183us over  400 runs of    85 symbols from emoji.txt                                   
    lex_new_grapheme_utf8 average time   192us over  400 runs of    85 symbols from emoji.txt DFA cleared        
```

After:

```
Warming up Java compiler....
    load_legacy_java_utf8 average time   247us over 3500 loads of 29038 symbols from Parser.java
    load_legacy_java_utf8 average time   294us over 3500 loads of 13379 symbols from udhr_hin.txt
            load_new_utf8 average time   754us over 3500 loads of 29038 symbols from Parser.java
            load_new_utf8 average time   573us over 3500 loads of 13379 symbols from udhr_hin.txt

     lex_legacy_java_utf8 average time   661us over 2000 runs of 29038 symbols
     lex_legacy_java_utf8 average time  1753us over 2000 runs of 29038 symbols DFA cleared
        lex_new_java_utf8 average time  1313us over 2000 runs of 29038 symbols
        lex_new_java_utf8 average time  2217us over 2000 runs of 29038 symbols DFA cleared

 lex_legacy_grapheme_utf8 average time 11915us over  400 runs of  6614 symbols from udhr_kor.txt
 lex_legacy_grapheme_utf8 average time 12356us over  400 runs of  6614 symbols from udhr_kor.txt DFA cleared
 lex_legacy_grapheme_utf8 average time 10346us over  400 runs of 13379 symbols from udhr_hin.txt
 lex_legacy_grapheme_utf8 average time 10506us over  400 runs of 13379 symbols from udhr_hin.txt DFA cleared
    lex_new_grapheme_utf8 average time 12087us over  400 runs of  6614 symbols from udhr_kor.txt
    lex_new_grapheme_utf8 average time 12174us over  400 runs of  6614 symbols from udhr_kor.txt DFA cleared
    lex_new_grapheme_utf8 average time 10762us over  400 runs of 13379 symbols from udhr_hin.txt
    lex_new_grapheme_utf8 average time 10873us over  400 runs of 13379 symbols from udhr_hin.txt DFA cleared
    lex_new_grapheme_utf8 average time   194us over  400 runs of    85 symbols from emoji.txt
    lex_new_grapheme_utf8 average time   204us over  400 runs of    85 symbols from emoji.txt DFA cleared
```

Now that I have the perf test running cleanly in my environment, I'll attach YourKit and see what I can fix.