@bhamiltoncx 2017-03-23
OK, with https://github.com/antlr/antlr4/pull/1788 on top of this:

```
    load_legacy_java_utf8 average time   272us size 132788b over 3500 loads of 29038 symbols from Parser.java
    load_legacy_java_utf8 average time   305us size 128671b over 3500 loads of 13379 symbols from udhr_hin.txt
            load_new_utf8 average time   745us size 116765b over 3500 loads of 29038 symbols from Parser.java
            load_new_utf8 average time   539us size 119317b over 3500 loads of 13379 symbols from udhr_hin.txt
```

I know how to improve the perf a few ways as well.