@parrt 2017-03-23
Hmm...load times seems slow (you got an SSD and fast clock)?  Also, the sizes are bigger not smaller for Parser.java.