@bhamiltoncx 2017-03-24
I suspect the memory usage logic and calls to GC are not actually doing what we want them to. Note the `-217b` in the "after" numbers below..

Last I checked, you can't actually force a GC in Java these days, the API just ignores what you ask it to do.

Oddly, I'm seeing similar numbers to you which suggest `lex_legacy_java_utf8` is affected by this diff stack. Like you, I have a hard time understanding how that's possible, none of that code should be touched here.

Before (master @ 00d5781bee53df06188554663b6edaddff51df3e):

```
   load_legacy_java_ascii average time   208us size  62584b over 3500 loads of 29038 symbols from Parser.java
    load_legacy_java_utf8 average time   197us size  35744b over 3500 loads of 29038 symbols from Parser.java
    load_legacy_java_utf8 average time   317us size   1033b over 3500 loads of 13379 symbols from udhr_hin.txt
            load_new_utf8 average time   472us size  98498b over 3500 loads of 29038 symbols from Parser.java
            load_new_utf8 average time   390us size  34815b over 3500 loads of 13379 symbols from udhr_hin.txt

     lex_legacy_java_utf8 average time   613us over 2000 runs of 29038 symbols
     lex_legacy_java_utf8 average time  1539us over 2000 runs of 29038 symbols DFA cleared
        lex_new_java_utf8 average time   648us over 2000 runs of 29038 symbols
        lex_new_java_utf8 average time  1580us over 2000 runs of 29038 symbols DFA cleared

 lex_legacy_grapheme_utf8 average time  6617us over  400 runs of  6614 symbols from udhr_kor.txt
 lex_legacy_grapheme_utf8 average time  6838us over  400 runs of  6614 symbols from udhr_kor.txt DFA cleared
 lex_legacy_grapheme_utf8 average time 12311us over  400 runs of 13379 symbols from udhr_hin.txt
 lex_legacy_grapheme_utf8 average time 12412us over  400 runs of 13379 symbols from udhr_hin.txt DFA cleared
    lex_new_grapheme_utf8 average time  6765us over  400 runs of  6614 symbols from udhr_kor.txt
    lex_new_grapheme_utf8 average time  6910us over  400 runs of  6614 symbols from udhr_kor.txt DFA cleared
    lex_new_grapheme_utf8 average time 12838us over  400 runs of 13379 symbols from udhr_hin.txt
    lex_new_grapheme_utf8 average time 12971us over  400 runs of 13379 symbols from udhr_hin.txt DFA cleared
    lex_new_grapheme_utf8 average time   148us over  400 runs of    85 symbols from emoji.txt
    lex_new_grapheme_utf8 average time   142us over  400 runs of    85 symbols from emoji.txt DFA cleared
```

After (9c93187fc58c58574de20b8b8f5e80e339c82d66):

```
% java -cp /data/users/beng/src/antlr4/runtime/Java/target/antlr4-runtime-4.7-SNAPSHOT.jar:/data/users/beng/src/antlr4/runtime-testsuite/target/antlr4-runtime-testsuite-4.7-SNAPSHOT.jar:/data/users/beng/src/antlr4/runtime-testsuite/target/antlr4-runtime-testsuite-4.7-SNAPSHOT-tests.jar org.antlr.v4.test.runtime.java.api.perf.TimeLexerSpeed
Java VM args: 
Warming up Java compiler....
   load_legacy_java_ascii average time   223us size  62572b over 3500 loads of 29038 symbols from Parser.java
    load_legacy_java_utf8 average time   214us size  35744b over 3500 loads of 29038 symbols from Parser.java
    load_legacy_java_utf8 average time   322us size   1332b over 3500 loads of 13379 symbols from udhr_hin.txt
            load_new_utf8 average time   214us size   -217b over 3500 loads of 29038 symbols from Parser.java
            load_new_utf8 average time   328us size  28449b over 3500 loads of 13379 symbols from udhr_hin.txt

     lex_legacy_java_utf8 average time  1025us over 2000 runs of 29038 symbols
     lex_legacy_java_utf8 average time  1955us over 2000 runs of 29038 symbols DFA cleared
        lex_new_java_utf8 average time  1007us over 2000 runs of 29038 symbols
        lex_new_java_utf8 average time  1822us over 2000 runs of 29038 symbols DFA cleared

 lex_legacy_grapheme_utf8 average time  7060us over  400 runs of  6614 symbols from udhr_kor.txt
 lex_legacy_grapheme_utf8 average time  7181us over  400 runs of  6614 symbols from udhr_kor.txt DFA cleared
 lex_legacy_grapheme_utf8 average time 13171us over  400 runs of 13379 symbols from udhr_hin.txt
 lex_legacy_grapheme_utf8 average time 13352us over  400 runs of 13379 symbols from udhr_hin.txt DFA cleared
    lex_new_grapheme_utf8 average time  7067us over  400 runs of  6614 symbols from udhr_kor.txt
    lex_new_grapheme_utf8 average time  7173us over  400 runs of  6614 symbols from udhr_kor.txt DFA cleared
    lex_new_grapheme_utf8 average time 13219us over  400 runs of 13379 symbols from udhr_hin.txt
    lex_new_grapheme_utf8 average time 13595us over  400 runs of 13379 symbols from udhr_hin.txt DFA cleared
    lex_new_grapheme_utf8 average time   143us over  400 runs of    85 symbols from emoji.txt
    lex_new_grapheme_utf8 average time   154us over  400 runs of    85 symbols from emoji.txt DFA cleared
```